Move traits and structures into mod
